Refactor(GameController.java): 메소드 이동 및 스캐너 주입
GameService에서 사용할 스캐너를 외부에서 주입해주는 것으로 변경: 미리 준비해둔 input data로 테스팅하기 위함
