+++ src/main/java/controller/GameController.java
@@ -1,5 +1,6 @@
 package controller;
 
+import java.util.Scanner;
 import model.Hints;
 import model.Number;
 import service.GameService;
@@ -8,21 +9,23 @@ import service.RandomGenerator;
 public class GameController {
 
     private final GameService gameService;
+    private final Scanner sc;
 
     public GameController() {
         this.gameService = new GameService(new RandomGenerator());
+        this.sc = new Scanner(System.in);
     }
 
     public void playGame() {
-
+        GameService.setScanner(sc);
         do {
-            playOneTurn();
-        } while (gameService.isContinuing());
+            Number answer = gameService.getAnswer();
+            playOneTurn(answer);
+        } while (isContinuing());
 
     }
 
-    public void playOneTurn() {
-        Number answer = gameService.getAnswer();
+    public void playOneTurn(Number answer) {
 
         while (true) {
             System.out.print("숫자를 입력해주세요: ");
@@ -33,11 +36,17 @@ public class GameController {
                 answer.getNumbers()); // result[0]: strike, result[1]: ball
 
             if (gameService.isAnswered(hints)) {
+                System.out.println("3개의 숫자를 모두 맞히셨습니다! 게임 종료");
                 break;
             } else {
                 gameService.printHints(hints);
             }
         }
 
+    }
+
+    public boolean isContinuing() {
+        System.out.println("게임을 새로 시작하려면 1, 종료하려면 2를 입력하세요.");
+        return gameService.getInput() == 1;
     }
 }
